docker007/requirements.txt
pandas
pandas_datareader
python-dotenv
SQLAlchemy
mysql-connector-python
apache-airflow
apache-airflow-providers-docker